requirements.txt
Pillow
opencv-python
numpy
imageio